requirements.txt
pandas
python-dotenv
pyyaml
requests
